+++ src/main/java/org/prgrms/vouchermanager/service/CustomerService.java
@@ -1,7 +1,7 @@
 package org.prgrms.vouchermanager.service;
 
 import lombok.RequiredArgsConstructor;
-import org.prgrms.vouchermanager.repository.CustomerRepositroy;
+import org.prgrms.vouchermanager.repository.customer.CustomerRepositroy;
 import org.prgrms.vouchermanager.domain.customer.Customer;
 import org.springframework.stereotype.Service;
