+++ src/test/java/org/prgrms/vouchermanager/repository/CsvCustomerRepositoryTest.java
@@ -2,6 +2,8 @@ package org.prgrms.vouchermanager.repository;
 
 import org.junit.jupiter.api.Test;
 import org.prgrms.vouchermanager.domain.customer.Customer;
+import org.prgrms.vouchermanager.repository.customer.CsvCustomerRepository;
+import org.prgrms.vouchermanager.repository.customer.CustomerRepositroy;
 
 import java.util.List;
